Move GameState enum to common module
